requirements.txt
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
streamlit==1.28.0
plotly==5.15.0
pandas==2.1.0